+++ cmd/tiltdemo/main.go
@@ -1,9 +1,17 @@
 package main
 
 import (
-	"github.com/windmilleng/tilt/internal/demo"
+	"fmt"
+	"time"
 )
 
 func main() {
-	demo.Execute()
+	println(0, "Starting tilt...")
+	println(100, "Building [my-app-frontend]")
+	println(500, "Build complete in 3.3423s")
+}
+
+func println(ms int, msg string) {
+	time.Sleep(time.Duration(ms) * time.Millisecond)
+	fmt.Println(msg)
 }
